refactor: move keymapping logic to dirs
