+++ init.lua
@@ -1,15 +1,8 @@
 require("config.lazy")
+require("keymaps")
 require("lsp")
 
 vim.cmd("colorscheme tokyonight-night")
-
-local cc_default = "80"
-vim.keymap.set("n", "<leader>c", function()
-  vim.api.nvim_buf_call(0, function()
-    vim.b["cc"] = vim.b["cc"] or cc_default
-    vim.wo["cc"] = (vim.wo["cc"] == vim.b["cc"]) and "" or vim.b["cc"]
-  end)
-end)
 
 vim.o.mouse = ""
 
@@ -21,15 +14,8 @@ vim.o.clipboard = "unnamedplus"
 vim.o.cursorline = true
 
 vim.o.hlsearch = true
-vim.keymap.set("n", "<Esc>", "<cmd>nohlsearch<CR>")
 
 vim.g.have_nerd_font = false
-
-vim.keymap.set("n", "<leader>j", "<c-w>", { silent = true })
-
-vim.keymap.set("n", "<leader>k", "<cmd>tabnew<CR>", { silent = true })
-
-vim.keymap.set("n", "<leader>h", "<cmd>Ex<CR>", { silent = true })
 
 vim.opt.list = true
 vim.opt.listchars = { tab = "» ", trail = "·", nbsp = "␣", leadmultispace = "↦ " }
@@ -162,51 +148,4 @@ for pattern, value in pairs(cc_values) do
     desc = string.format("Initalize colorcolumn value for pattern: %s.", pattern)
   })
 end
-
-local Terminal = require("toggleterm.terminal").Terminal
-
-function _on_open(term)
-  vim.cmd("startinsert!")
-  local keymap_opts = { noremap = true, silent = true, buffer = term.bufnr }
-  vim.keymap.set("t", "<c-]>", "<c-\\><c-n>", keymap_opts)
-  vim.keymap.set("n", "q", "<cmd>close<CR>", keymap_opts)
-end
-
-function _on_close(term)
-  vim.cmd("startinsert!")
-end
-
-local lazygit = Terminal:new({
-  cmd = "lazygit",
-  hidden = true,
-  on_open = _on_open,
-  on_close = _on_close,
-})
-local tmux_cwd = Terminal:new({
-  cmd = string.format("tmux new-session -A -s %s", vim.fn.getcwd()),
-  hidden = true,
-  on_open = _on_open,
-  on_close = _on_close,
-})
-
-function _lazygit_toggle()
-  lazygit:toggle()
-end
-
-function _tmux_cwd_toggle()
-  tmux_cwd:toggle()
-end
-
-vim.keymap.set(
-  "n",
-  "<leader>lh",
-  "<cmd>lua _lazygit_toggle()<CR>",
-  { noremap = true, silent = true }
-)
-vim.keymap.set(
-  "n",
-  "<leader>lj",
-  "<cmd>lua _tmux_cwd_toggle()<CR>",
-  { noremap = true, silent = true }
-)
